Move more data types to the data folder.
