+++ Vision/Module.cs
@@ -1,6 +1,6 @@
 ﻿global using Detection = Tyr.Common.Data.Ssl.Vision.Detection;
-global using RawBall = Tyr.Vision.RawDetection<Tyr.Common.Data.Ssl.Vision.Detection.Ball>;
-global using RawRobot = Tyr.Vision.RawDetection<Tyr.Common.Data.Ssl.Vision.Detection.Robot>;
+global using RawBall = Tyr.Vision.Data.RawDetection<Tyr.Common.Data.Ssl.Vision.Detection.Ball>;
+global using RawRobot = Tyr.Vision.Data.RawDetection<Tyr.Common.Data.Ssl.Vision.Detection.Robot>;
 using Tyr.SourceGen;
 
 namespace Tyr.Vision;
